move checkParameters to separate file
