+++ package.json
@@ -7,7 +7,8 @@
     "async-shim.js",
     "browser.js",
     "index.js",
-    "node-shim.js"
+    "node-shim.js",
+    "precondition.js"
   ],
   "browser": "./browser.js",
   "keywords": [
